emprestapi/requirements.txt
django
djangorestframework
psycopg2
django-rest-auth
django-allauth